requirements.txt
scikit-learn
scipy
numpy
pandas
requests
tornado
slackclient
pytimeparse
PyYAML
